+++ CHANGES.md
@@ -4,6 +4,7 @@ Version 0.8.5 (pending)
 - Fixed pronouns in staff import
 - Improved signature pad detection
 - Fixed redirect to non-existent page when printing staff badges
+- Mark guests as checked in during attendee import
 
 Version 0.8.4 (10/26/2019)
 - Added attendance report